Split file to enable lazy-load
